Uses RawIO or StimfitIO to read dat adn plot it
I have used raw io to open and visualize the data but got stuck at the point where it's trying to plot and some of the data doesn't have right dimensions.  Or it is reading quite differently than what I get using slimfitIO. StimfitIO reads and behaves nicely but only available in python2.